Add MapDraggable class Change Room and MoveChange to use MapDraggable
